pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.23.0",
    "aiohttp>=3.13.2",
    "dnspython>=2.8.0",
    "python-dotenv>=1.2.1",
    "telethon>=1.42.0",
    "cryptography>=42.0.0",
    "reportlab>=4.0.0",
    "jinja2>=3.1.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "openai>=1.0.0",
]
